clustering/requirements.txt
numpy
panda
Pillow==9.5.0
opencv-python
torch
torchvision
scikit-learn
